refactor: the direct-XML writers get their own file, and the rest of the list
Closes the remaining audit items. The big one is a pure move; the rest are
small and each one stops a result from being read as something it is not.

modifyD365File.ts: 6,090 -> 4,401 lines
  The audit found it at 5,645 and growing 44% between audits — the argument
  schema, a 62-arm operation switch, the disk locators and fifteen
  direct-XML writers in one file. The writers moved to
  directXmlWriters.ts. It is a clean cut, not a rearrangement: the block
  declared 35 symbols, exported the 19 the switch consumes, and needed
  exactly ONE from its old home (`serializedOnFile`, used nowhere else in
  src, so it came too). Behaviour is unchanged and the suite is the proof —
  plus a live write probe on the VM, since these writers ARE the path a
  create-with-operations, an add-field, a table-extension create and a
  delete take. layering.test.ts now pins the size so the split cannot
  silently undo itself.

Bridge
  • A health-check ping that misses while a WRITE is outstanding no longer
    restarts the child. Writes take every read slot by design, so nothing —
    ping included — is answered while one runs; the 5-second ping timeout
    therefore read a busy child as a wedged one and killed it mid-write.
    An AOT write is not transactional. Busy is not wedged.
  • "0 parent duties" now says whether it found none or could not finish.
    ReadSecurityPrivilege scans every duty and swallowed a failure into an
    empty list, which the renderer stated as fact — and "this privilege is
    in no duty" is exactly the claim BPErrorPrivilegeNotCoveredByDuty is
    about, acted on by adding it to a duty it may already be in.
  • The head-of-line serialisation was NOT changed, deliberately: a write
    holding every read slot is what stops it overlapping a read of the
    provider it replaces. Moving it off the loop thread would let the loop
    parse the next request and no more, since reads still cannot run. The
    restart guard above is the part that was actually costing something.

Smaller, same theme
  • `dbSync: []` synced everything — an empty list fell through with no
    `tables`, which dbSyncTool reads as "derive the scope from the project".
    Asking to sync nothing is a mistake, not an instruction.
  • A failed build kept only a 40-line tail whenever ANY diagnostic parsed,
    warnings included — so a failure whose error shape the regexes miss, in
    a log carrying BP warnings, lost the log that held the reason. The tail
    is for when an ERROR explains the failure.
  • The once-per-session "no projectPath" key was the candidate COUNT, so
    two workspaces with the same number of .rnrproj files shared one key and
    the second never saw its own list. Keyed by the candidates now.
  • indexStaleness's background-scan test slept a fixed 20 ms — comfortable
    on an idle dev box, a flake on a loaded runner. Polls to a deadline.

Not changed, with reasons: the schema budget's 81 chars of headroom is the
repo's stated policy ("headroom is small on purpose so creep is caught
early"), and the class-authoring guidance dropped from d365fo_file's tool
description still reads verbatim on `sourceCode`, the parameter it is about.

Full suite 5,008 passing. Bridge re-attested.
